refactor(test): migrate slack.ts helpers to three-tier structure
Move all DB-direct helpers from api-test-helpers/slack.ts into:
- db-test-seeders/slack.ts: 7 seeder functions with @why-db-direct docs
- db-test-assertions/slack.ts: 7 read-only assertion/find/count functions

Delete api-test-helpers/slack.ts entirely and remove barrel re-export.
Update SLACK_BOT_SCOPES to import directly from source.
Update imports across 17 consumer test files.

Closes #9351
