+++ turbo/apps/web/src/lib/zero/slack-org/__tests__/connect-service.test.ts
@@ -7,9 +7,11 @@ import {
 import {
   createTestSlackOrgInstallation,
   seedTestSlackOrgConnection,
+} from "../../../../__tests__/db-test-seeders/slack";
+import {
   countSlackOrgInstallations,
   countSlackOrgConnections,
-} from "../../../../__tests__/api-test-helpers";
+} from "../../../../__tests__/db-test-assertions/slack";
 import {
   adminConnect,
   memberConnect,
