+++ turbo/apps/web/src/lib/zero/user/__tests__/user-deletion-service.test.ts
@@ -12,9 +12,6 @@ import {
   findTestRunRecord,
   findTestQueueEntry,
   insertTestQueueEntry,
-  insertTestSlackOrgInstallation,
-  insertTestSlackOrgConnection,
-  insertTestSlackOrgThreadSession,
   insertTestCreditUsageForRun,
   insertTestConversation,
   insertTestStorage,
@@ -25,7 +22,6 @@ import {
   insertOrgMembersEntry,
   insertUserCacheEntry,
   countUserRows,
-  countSlackConnectionRows,
   countGithubUserLinkRows,
   countTelegramUserLinkRows,
   insertTestGithubInstallation,
@@ -33,6 +29,12 @@ import {
   insertTestTelegramInstallation,
   insertTestTelegramUserLink,
 } from "../../../../__tests__/api-test-helpers";
+import {
+  insertTestSlackOrgInstallation,
+  insertTestSlackOrgConnection,
+  insertTestSlackOrgThreadSession,
+} from "../../../../__tests__/db-test-seeders/slack";
+import { countSlackConnectionRows } from "../../../../__tests__/db-test-assertions/slack";
 import { insertTestComposeJob } from "../../../../__tests__/db-test-seeders/agents";
 import { deleteUserData } from "../user-deletion-service";
 import { seedTestRun } from "../../../../__tests__/db-test-seeders/runs";
